backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework-simplejwt==5.5.1
redis==6.4.0
celery==5.5.3
django-celery-beat==2.8.1
drf-yasg==1.21.9
